# deflate: Speed up huffman encoding