Huffman only:

320.36 -> 414.67MB/s  (github-june-2days-2019.json)
237.41-> 293.24 MB/s (github-ranks-backup.bin)

Level 1:

278.94 -> 308.47 MB/s  (github-june-2days-2019.json)
232.93 -> 261.16MB/s  (github-ranks-backup.bin)